requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
transformers==4.43.0
torch>=1.12.0
diffusers==0.18.0
accelerate==0.20.3
Pillow==9.5.0
streamlit==1.26.0
boto3==1.28.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.4
python-multipart==0.0.6
# Note: diffusers + accelerate + torch can be heavy. If you plan to use the Hugging Face Inference API instead, you can omit diffusers/accelerate and use requests to call the API.
